+++ demo/src/types.d.ts
@@ -1,6 +1,6 @@
 declare module '@tensil/kinetic-input' {
-  export const CollapsibleNumberPicker: any;
-  export const StandaloneWheelPicker: any;
+  export const CollapsiblePicker: any;
+  export const Picker: any;
   export const PickerGroup: any;
   export const PickerColumn: any;
   export const PickerItem: any;